+++ multisig/simulations/ethereum/vip-024/abi/PoolRegistryAbi.json
