+++ multisig/simulations/ethereum/vip-024/abi/ResilientOracleAbi.json
