+++ multisig/simulations/ethereum/vip-024/abi/VTokenAbi.json
