+++ multisig/simulations/sepolia/index.ts
@@ -11,4 +11,4 @@ import "./vip-009";
 import "./vip-013";
 import "./vip-016";
 import "./vip-018";
-import "./vip-019";
+import "./vip-024";
