+++ multisig/simulations/sepolia/vip-024/abi/PoolRegistryAbi.json
