+++ multisig/simulations/sepolia/vip-024/abi/ResilientOracleAbi.json
