+++ multisig/simulations/sepolia/vip-024/index.ts
@@ -8,8 +8,8 @@ import { checkIsolatedPoolsComptrollers } from "../../../../src/vip-framework/ch
 import { checkVToken } from "../../../../src/vip-framework/checks/checkVToken";
 import { checkInterestRate } from "../../../../src/vip-framework/checks/interestRateModel";
 import { forking, pretendExecutingVip } from "../../../../src/vip-framework/index";
-import vip019 from "../../../proposals/sepolia/vip-019";
-import { BORROW_CAP, SUPPLY_CAP, TUSD, vTUSD } from "../../../proposals/sepolia/vip-019";
+import vip024 from "../../../proposals/sepolia/vip-024";
+import { BORROW_CAP, SUPPLY_CAP, TUSD, vTUSD } from "../../../proposals/sepolia/vip-024";
 import COMPTROLLER_ABI from "./abi/ComptrollerAbi.json";
 import POOL_REGISTRY_ABI from "./abi/PoolRegistryAbi.json";
 import RESILIENT_ORACLE_ABI from "./abi/ResilientOracleAbi.json";
@@ -44,7 +44,7 @@ forking(5744100, () => {
 
   describe("Post-VIP behavior", async () => {
     before(async () => {
-      await pretendExecutingVip(vip019());
+      await pretendExecutingVip(vip024());
     });
 
     it("check price", async () => {
