+++ vips/vip-296/bscmainnet.ts
@@ -9,10 +9,10 @@ const { bscmainnet } = NETWORK_ADDRESSES;
 export const USDT = "0x55d398326f99059fF775485246999027B3197955";
 export const COMMUNITY_WALLET = "0xc444949e0054A23c44Fc45789738bdF64aed2391";
 
-export const vip291 = () => {
+export const vip296 = () => {
   const meta = {
     version: "v2",
-    title: "VIP-291 Ethereum: add TUSD market to core pool",
+    title: "VIP-296 Ethereum: add TUSD market to core pool",
     description: ``,
     forDescription: "I agree that Venus Protocol should proceed with this proposal",
     againstDescription: "I do not think that Venus Protocol should proceed with this proposal",
@@ -32,4 +32,4 @@ export const vip291 = () => {
   );
 };
 
-export default vip291;
+export default vip296;
